+++ netsgiro/records.py
@@ -552,7 +552,7 @@ class TransactionAmountItem2(TransactionRecord):
             ).format(self=self)
         elif self.service_code == netsgiro.ServiceCode.AVTALEGIRO:
             service_fields = (
-                '{self.payer_name:10}'
+                '{:10}'.format(self.payer_name[:10])
                 + (' ' * 25)
                 + (self.reference and '{self.reference:25}' or (' ' * 25))
                 + ('0' * 5)